requirements.txt
accelerate
imagehash
pillow
torch
torchvision
transformers
